Add tests for async_setup_entry
